server: move InnerInMemory into a separate module
